requirements.txt
openai==0.27.8
pandas==1.4.3
requests
scikit-learn==1.1.1
sentence_transformers==2.7.0
tqdm
faiss-cpu==1.7.2
numpy==1.21.0
streamlit==1.23.1
notebook
